format code and move GoogleUtil to util
